Create Redshift
In this assignment, you were tasked with setting up an Amazon Redshift data warehouse for XYZ Corporation’s application. The goal was to implement a solution that could store and retrieve large datasets efficiently, enabling advanced data analytics.

Tasks Performed:
Create a Redshift Data Warehouse:

The first step was to set up an Amazon Redshift cluster to serve as the central data warehouse. This involved configuring the necessary resources and ensuring the Redshift cluster was ready for use.

Using the Query Editor:

Load Data: The next task was to upload data into the Redshift data warehouse. This process utilized the query editor to insert data into the warehouse for future analysis.

Query the Data: After loading the data, you used the query editor to execute SQL queries to analyze and retrieve the stored data, simulating real-world usage of Redshift for data analytics.